refactor git-diff execution list filter
